requirements.txt
asttokens==2.4.0
backcall==0.2.0
certifi==2020.6.20
chardet==4.0.0
colorzero==1.1
dbus-python==1.2.16
decorator==5.1.1
distro==1.5.0
exceptiongroup==1.1.3
executing==1.2.0
gpiozero==1.6.2
idna==2.10
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
numpy==1.25.2
opencv-python-headless==4.8.0.76
parso==0.8.3
pexpect==4.8.0
picamera2==0.3.12
pickleshare==0.7.5
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyamapping==0.1.0
Pygments==2.16.1
PyGObject==3.38.0
python-apt==2.2.1
python-osc==1.8.3
python-prctl==1.7
requests==2.25.1
RPi.GPIO==0.7.0
sc3nb==1.1.0
scipy==1.11.2
simplejpeg==1.6.4
six==1.16.0
spidev==3.5
ssh-import-id==5.10
ST7789==0.0.4
stack-data==0.6.2
toml==0.10.1
traitlets==5.9.0
typing-extensions==4.7.1
urllib3==1.26.5
v4l2-python3==0.3.2
wcwidth==0.2.6
